+++ sulong/projects/com.oracle.truffle.llvm.runtime/src/com/oracle/truffle/llvm/runtime/nodes/memory/LLVMVarArgCompoundAddressNode.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, 2019, Oracle and/or its affiliates.
+ * Copyright (c) 2017, 2020, Oracle and/or its affiliates.
  *
  * All rights reserved.
  *
@@ -37,9 +37,9 @@ import com.oracle.truffle.llvm.runtime.LLVMVarArgCompoundValue;
 import com.oracle.truffle.llvm.runtime.nodes.api.LLVMExpressionNode;
 
 @NodeChild(type = LLVMExpressionNode.class, value = "source")
-@NodeFields({@NodeField(name = "length", type = int.class), @NodeField(name = "alignment", type = int.class)})
+@NodeFields({@NodeField(name = "length", type = long.class), @NodeField(name = "alignment", type = int.class)})
 public abstract class LLVMVarArgCompoundAddressNode extends LLVMExpressionNode {
-    public abstract int getLength();
+    public abstract long getLength();
 
     public abstract int getAlignment();
